Well if I am not stupid, this should work. Dunno what is gonna happen when I click check-in since all these changes were done last night and the night before (9/26 -28/11).
What's New:

SPLASHSCREEN!
Since parsing at start takes a bit of time, I figured, for now I would try to give people something to look at :D (the pic is from the greendragon fight (Timmy) on deli and where I got my Dragonslayer title XD )

Dynamicly changing users! (with refresh)
It annoyed me that the tree was empty on first use, so I changed that! with that you can now select which user, GUI isn't finished so that placement isn't going to stay that way.

O.O! *nervously clicks check-in button and waites for explosion.*

I have started work on changing maybe some of WParse to incorporate skill gain over a time span, this is still being worked on and with that somethings are out of place in this set is ment to be that way as I am simply trying to get my builds and commits into CP.