Added the start of my custom DLL
Made Splashscreen part of that dll.

refactored some code

Fixed an oddity when RegUser had cut off the first letter in the user name. It was because there was an actual differnce in the registry value..the / was gone! So we check now.

Started commenting the code. Really need to do this before anything else (for my own reasons at least)

Settings can no longer be runtime launched. Forms there but no code points to it.

Removed Form1, going another way with idea :)

Fixed a stupid thing I did. Made it impossible to switch users with the combo box. This is what happens when your under the influeance :P

Starting remembering how to performance test methods and such, gonne need that!

Gave AggregateParser another overload. Prep for skillz...